+++ R/OptimInstanceMultiCrit.R
@@ -25,8 +25,12 @@ OptimInstanceMultiCrit = R6Class("OptimInstanceMultiCrit",
     #' @param objective ([Objective]).
     #' @param terminator ([Terminator])\cr
     #' Multi-criteria terminator.
-    initialize = function(objective, search_space = NULL, terminator) {
-      super$initialize(objective, search_space, terminator)
+    #' @param check_values (`logical(1)`)\cr
+    #' Should x-values that are added to the archive be checked for validity?
+    #' Search space that is logged into archive.
+    initialize = function(objective, search_space = NULL, terminator,
+      check_values = TRUE) {
+      super$initialize(objective, search_space, terminator, check_values)
     },
 
     #' @description